requirements.txt
fastapi
uvicorn[standard]
pytesseract 
pillow 
tesseract
python-multipart
python-dotenv